requirements.txt
fastapi
uvicorn
python-dotenv
google-cloud-speech 
google-cloud-texttospeech 
google-cloud-translate
google-cloud-firestore
pytest
pydantic-settings
pytest-asyncio
google-cloud-aiplatform
